+++ CHANGES.md
@@ -1,3 +1,6 @@
+- v0.9.0
+    - `unikmer`: new binary format with support for taxid.
+    - deleted command: `unikmer subset`.
 - v0.8.0
     - `unikmer`:
          - new option `-i/--infile-list`, if given, files in the list file are appended to files from cli arguments.